requirements.txt
pydub
librosa
SpeechRecognition
ffmpeg-python
opencv-python
paddlepaddle-gpu
paddleocr
matplotlib
gtts
pillow
pytesseract
sounddevice
moviepy
numpy
pyaudio
